+++ CSharpCodeAnalyst/Gallery/Gallery.cs
@@ -5,11 +5,6 @@ namespace CSharpCodeAnalyst.Gallery;
 [Serializable]
 public class Gallery
 {
-    public Gallery()
-    {
-
-    }
-
     public List<GraphSessionState> Sessions { get; set; } = new();
 
     public void AddSession(GraphSessionState session)
