+++ Contracts/Graph/CodeElement.cs
@@ -8,8 +8,8 @@ public class CodeElement(string id, CodeElementType elementType, string name, st
     public List<SourceLocation> SourceLocations { get; set; } = [];
 
     /// <summary>
-    /// Unlike in the dependency graph where external dependencies are omitted
-    /// I want to keep all attributes here.
+    ///     Unlike in the dependency graph where external dependencies are omitted
+    ///     I want to keep all attributes here.
     /// </summary>
     public HashSet<string> Attributes { get; set; } = [];
